Move compose files to lib under compose/
No functional changes, simply moving all compose code under this
directory in preparation for building libcompose.

libsend depends on libcompose, so set it first under MUTTLIBS in
Makefile.autosetup

Move compose config to composelib
