+++ send/send.c
@@ -46,9 +46,9 @@
 #include "mutt.h"
 #include "send.h"
 #include "lib.h"
+#include "compose/lib.h"
 #include "ncrypt/lib.h"
 #include "pattern/lib.h"
-#include "compose.h"
 #include "context.h"
 #include "copy.h"
 #include "handler.h"
